+++ core/js/src/index.ts
@@ -3,3 +3,4 @@ export * from "./merge_row_batch";
 export * from "./score";
 export * from "./util";
 export * from "./span_types";
+export * from "./git_fields";
